refactor(codegen): extrai compile/mir_route.rs de func.rs
Move clear_mir_cache_for_program + try_compile_via_mir + debug_bail
+ mir_param_compatible + mir_ret_compatible (~294 linhas) e
o thread_local MIR_CACHE para compile/mir_route.rs.

UserFn struct fica em func.rs mas agora pub(crate) com fields
pub(crate) para que mir_route possa indexar.

func.rs: 1453 -> 1167 linhas. Suite verde.
